complete password support
