+++ package.json
@@ -10,12 +10,12 @@
   "scripts": {
     "android": "yarn run build-rust-android && npx react-native run-android",
     "build-prod:ios": "NODE_OPTIONS=--max_old_space_size=8192 npx react-native bundle --dev false --entry-file index.js --bundle-output ios/main.jsbundle --platform ios --assets-dest='./ios'",
-    "build-rust-ios": "cd rust/signer && make ios",
-    "build:ios": "yarn install && cd ios && pod install && cd .. && yarn build-rust-ios",
+    "build-rust-ios": "(cd rust/signer && make ios)",
+    "build:ios": "yarn install && (cd ios && pod install) && yarn build-rust-ios",
     "build:android": "yarn install && yarn buil-rust-android",
-    "build-rust-android": "cd rust/signer && make android",
+    "build-rust-android": "(cd rust/signer && make android)",
     "clean": "watchman watch-del-all && rm -rf /tmp/metro-bundler-cache-* && rm -rf /tmp/haste-map-react-native-packager-* && rm -rf node_modules/ && yarn cache clean --force && rm -rf ./android/app/build && rm -rf ./ios/build && yarn clean-rust && cd ios && pod deintegrate && cd ..",
-    "clean-rust": "cd rust/signer && cargo clean && cd ../..",
+    "clean-rust": "(cd rust/signer && cargo clean)",
     "commit": "commit-wizard",
     "ios": "yarn run build-rust-ios && npx react-native run-ios",
     "lint": "npx eslint . --ext .js,.jsx,.ts,.tsx --ignore-path .gitignore",
@@ -23,7 +23,7 @@
     "start": "NODE_OPTIONS=--max_old_space_size=8192 npx react-native start",
     "unit": "jest --config ./test/unit/jest.config.js",
     "unit:debug": "node --inspect node_modules/.bin/jest --watch --runInBand",
-    "test-rust": "cd ./rust/signer && cargo test && cd ../..",
+    "test-rust": "(cd ./rust/signer && cargo test)",
     "build-e2e:android": "detox build -c android.emu.debug -l info",
     "test-e2e:android": "detox test -c android.emu.debug -l info --noStackTrace",
     "e2e:android": "yarn run build-e2e:android && yarn run test-e2e:android",
