+++ src/modules/Unlock/utils.ts
@@ -1,0 +1,44 @@
+// Copyright 2015-2020 Parity Technologies (UK) Ltd.
+// This file is part of Parity.
+
+// Parity is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+// Parity is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+// You should have received a copy of the GNU General Public License
+// along with Parity.  If not, see <http://www.gnu.org/licenses/>.
+
+import t from 'modules/unlock/strings';
+import { State, UpdateStateFunc } from 'modules/unlock/types';
+import { onlyNumberRegex } from 'utils/regex';
+
+type InputListener = (v: string) => void;
+
+export const onPinInputChange = (
+	stateName: 'pin' | 'confirmation',
+	updateState: UpdateStateFunc
+): InputListener => (pinInput: string) => {
+	if (onlyNumberRegex.test(pinInput)) {
+		updateState({
+			pinMismatch: false,
+			pinTooShort: false,
+			[stateName]: pinInput
+		});
+	}
+};
+
+export const getSubtitle = (state: State, isUnlock: boolean): string => {
+	if (state.pinTooShort) {
+		return t.pinTooShortHint;
+	} else if (state.pinMismatch) {
+		return isUnlock
+			? t.pinMisMatchHint.pinUnlock
+			: t.pinMisMatchHint.pinCreation;
+	}
+	return isUnlock ? t.subtitle.pinUnlock : t.subtitle.pinCreation;
+};
