+++ src/screens.tsx
@@ -40,7 +40,6 @@ import { AccountUnlock, AccountUnlockAndSign } from 'screens/AccountUnlock';
 import IdentityBackup from 'screens/IdentityBackup';
 import IdentityManagement from 'screens/IdentityManagement';
 import IdentityNew from 'screens/IdentityNew';
-import IdentityPin from 'screens/IdentityPin';
 import LegacyAccountBackup from 'screens/LegacyAccountBackup';
 import LegacyAccountList from 'screens/LegacyAccountList';
 import LegacyNetworkChooser from 'screens/LegacyNetworkChooser';
@@ -49,6 +48,9 @@ import PathDerivation from 'screens/PathDerivation';
 import PathDetails from 'screens/PathDetails';
 import PathManagement from 'screens/PathManagement';
 import PathsList from 'screens/PathsList';
+import PinNew from 'modules/unlock/screens/PinNew';
+import PinUnlock from 'modules/unlock/screens/PinUnlock';
+import PinUnlockWithPassword from 'modules/unlock/screens/PinUnlockWithPassword';
 import PrivacyPolicy from 'screens/PrivacyPolicy';
 import QrScanner from 'screens/QrScanner';
 import Security from 'screens/Security';
@@ -144,12 +146,17 @@ export const AppNavigator = (): React.ReactElement => (
 			component={IdentityManagement}
 		/>
 		<ScreenStack.Screen name="IdentityNew" component={IdentityNew} />
-		<ScreenStack.Screen name="IdentityPin" component={IdentityPin} />
 		<ScreenStack.Screen name="MessageDetails" component={MessageDetails} />
 		<ScreenStack.Screen name="PathDerivation" component={PathDerivation} />
 		<ScreenStack.Screen name="PathDetails" component={PathDetails} />
 		<ScreenStack.Screen name="PathsList" component={PathsList} />
 		<ScreenStack.Screen name="PathManagement" component={PathManagement} />
+		<ScreenStack.Screen name="PinNew" component={PinNew} />
+		<ScreenStack.Screen name="PinUnlock" component={PinUnlock} />
+		<ScreenStack.Screen
+			name="PinUnlockWithPassword"
+			component={PinUnlockWithPassword}
+		/>
 		<ScreenStack.Screen name="QrScanner" component={QrScanner} />
 		<ScreenStack.Screen name="Security" component={Security} />
 		<ScreenStack.Screen name="SignedMessage" component={SignedMessage} />
