+++ src/utils/identitiesUtils.ts
@@ -15,8 +15,8 @@
 // along with Parity.  If not, see <http://www.gnu.org/licenses/>.
 
 import { pathsRegex } from './regex';
-import { decryptData } from './native';
-import { parseSURI } from './suri';
+import { decryptData, substrateAddress } from './native';
+import { constructSURI, parseSURI } from './suri';
 import { generateAccountId } from './account';
 
 import {
@@ -233,6 +233,24 @@ export const unlockIdentitySeed = async (
 	const seed = await decryptData(encryptedSeed, pin);
 	const { phrase } = parseSURI(seed);
 	return phrase;
+};
+
+export const verifyPassword = async (
+	password: string,
+	seedPhrase: string,
+	identity: Identity,
+	path: string
+): Promise<boolean> => {
+	const suri = constructSURI({
+		derivePath: path,
+		password: password,
+		phrase: seedPhrase
+	});
+	const networkKey = getNetworkKey(path, identity);
+	const networkParams = SUBSTRATE_NETWORK_LIST[networkKey];
+	const address = await substrateAddress(suri, networkParams.prefix);
+	const accountMeta = identity.meta.get(path);
+	return address === accountMeta?.address;
 };
 
 export const getExistedNetworkKeys = (identity: Identity): string[] => {
